Split test_advanced_n.py and enabled cluster tests
