+++ python/ray/tests/BUILD
@@ -31,6 +31,11 @@ py_test_module_list(
     "test_advanced_2.py",
     "test_advanced_3.py",
     "test_advanced_4.py",
+    "test_advanced_5.py",
+    "test_advanced_6.py",
+    "test_advanced_7.py",
+    "test_advanced_8.py",
+    "test_advanced_9.py",
     "test_array.py",
     "test_autoscaling_policy.py",
     "test_basic.py",
