+++ python/ray/tests/test_advanced.py
@@ -1,10 +1,7 @@
 # coding: utf-8
-from concurrent.futures import ThreadPoolExecutor
 import json
 import logging
-import random
 import sys
-import threading
 import time
 
 import os
@@ -17,7 +14,6 @@ import ray._private.profiling as profiling
 from ray._private.test_utils import (
     client_test_enabled,
     RayTestTimeoutException,
-    SignalActor,
 )
 from ray.exceptions import ReferenceCountingAssertionError
 
@@ -234,8 +230,8 @@ def test_profiling_api(ray_start_2_cpus):
         time.sleep(1.1)
 
 
-def test_wait_cluster(ray_start_cluster):
-    cluster = ray_start_cluster
+def test_wait_cluster(ray_start_cluster_enabled):
+    cluster = ray_start_cluster_enabled
     cluster.add_node(num_cpus=1, resources={"RemoteResource": 1})
     cluster.add_node(num_cpus=1, resources={"RemoteResource": 1})
     ray.init(address=cluster.address)
@@ -261,8 +257,8 @@ def test_wait_cluster(ray_start_cluster):
 
 
 @pytest.mark.skip(reason="TODO(ekl)")
-def test_object_transfer_dump(ray_start_cluster):
-    cluster = ray_start_cluster
+def test_object_transfer_dump(ray_start_cluster_enabled):
+    cluster = ray_start_cluster_enabled
 
     num_nodes = 3
     for i in range(num_nodes):
@@ -389,457 +385,6 @@ def test_illegal_api_calls(ray_start_regular):
     # Verify that we cannot call get on a regular value.
     with pytest.raises(Exception):
         ray.get(3)
-
-
-@pytest.mark.skipif(
-    client_test_enabled(), reason="grpc interaction with releasing resources"
-)
-def test_multithreading(ray_start_2_cpus):
-    # This test requires at least 2 CPUs to finish since the worker does not
-    # release resources when joining the threads.
-
-    def run_test_in_multi_threads(test_case, num_threads=10, num_repeats=25):
-        """A helper function that runs test cases in multiple threads."""
-
-        def wrapper():
-            for _ in range(num_repeats):
-                test_case()
-                time.sleep(random.randint(0, 10) / 1000.0)
-            return "ok"
-
-        executor = ThreadPoolExecutor(max_workers=num_threads)
-        futures = [executor.submit(wrapper) for _ in range(num_threads)]
-        for future in futures:
-            assert future.result() == "ok"
-
-    @ray.remote
-    def echo(value, delay_ms=0):
-        if delay_ms > 0:
-            time.sleep(delay_ms / 1000.0)
-        return value
-
-    def test_api_in_multi_threads():
-        """Test using Ray api in multiple threads."""
-
-        @ray.remote
-        class Echo:
-            def echo(self, value):
-                return value
-
-        # Test calling remote functions in multiple threads.
-        def test_remote_call():
-            value = random.randint(0, 1000000)
-            result = ray.get(echo.remote(value))
-            assert value == result
-
-        run_test_in_multi_threads(test_remote_call)
-
-        # Test multiple threads calling one actor.
-        actor = Echo.remote()
-
-        def test_call_actor():
-            value = random.randint(0, 1000000)
-            result = ray.get(actor.echo.remote(value))
-            assert value == result
-
-        run_test_in_multi_threads(test_call_actor)
-
-        # Test put and get.
-        def test_put_and_get():
-            value = random.randint(0, 1000000)
-            result = ray.get(ray.put(value))
-            assert value == result
-
-        run_test_in_multi_threads(test_put_and_get)
-
-        # Test multiple threads waiting for objects.
-        num_wait_objects = 10
-        objects = [echo.remote(i, delay_ms=10) for i in range(num_wait_objects)]
-
-        def test_wait():
-            ready, _ = ray.wait(
-                objects,
-                num_returns=len(objects),
-                timeout=1000.0,
-            )
-            assert len(ready) == num_wait_objects
-            assert ray.get(ready) == list(range(num_wait_objects))
-
-        run_test_in_multi_threads(test_wait, num_repeats=1)
-
-    # Run tests in a driver.
-    test_api_in_multi_threads()
-
-    # Run tests in a worker.
-    @ray.remote
-    def run_tests_in_worker():
-        test_api_in_multi_threads()
-        return "ok"
-
-    assert ray.get(run_tests_in_worker.remote()) == "ok"
-
-    # Test actor that runs background threads.
-    @ray.remote
-    class MultithreadedActor:
-        def __init__(self):
-            self.lock = threading.Lock()
-            self.thread_results = []
-
-        def background_thread(self, wait_objects):
-            try:
-                # Test wait
-                ready, _ = ray.wait(
-                    wait_objects,
-                    num_returns=len(wait_objects),
-                    timeout=1000.0,
-                )
-                assert len(ready) == len(wait_objects)
-                for _ in range(20):
-                    num = 10
-                    # Test remote call
-                    results = [echo.remote(i) for i in range(num)]
-                    assert ray.get(results) == list(range(num))
-                    # Test put and get
-                    objects = [ray.put(i) for i in range(num)]
-                    assert ray.get(objects) == list(range(num))
-                    time.sleep(random.randint(0, 10) / 1000.0)
-            except Exception as e:
-                with self.lock:
-                    self.thread_results.append(e)
-            else:
-                with self.lock:
-                    self.thread_results.append("ok")
-
-        def spawn(self):
-            wait_objects = [echo.remote(i, delay_ms=10) for i in range(10)]
-            self.threads = [
-                threading.Thread(target=self.background_thread, args=(wait_objects,))
-                for _ in range(20)
-            ]
-            [thread.start() for thread in self.threads]
-
-        def join(self):
-            [thread.join() for thread in self.threads]
-            assert self.thread_results == ["ok"] * len(self.threads)
-            return "ok"
-
-    actor = MultithreadedActor.remote()
-    actor.spawn.remote()
-    ray.get(actor.join.remote()) == "ok"
-
-
-@pytest.mark.skipif(client_test_enabled(), reason="internal api")
-def test_wait_makes_object_local(ray_start_cluster):
-    cluster = ray_start_cluster
-    cluster.add_node(num_cpus=0)
-    cluster.add_node(num_cpus=2)
-    ray.init(address=cluster.address)
-
-    @ray.remote
-    class Foo:
-        def method(self):
-            return np.zeros(1024 * 1024)
-
-    a = Foo.remote()
-
-    # Test get makes the object local.
-    x_id = a.method.remote()
-    assert not ray.worker.global_worker.core_worker.object_exists(x_id)
-    ray.get(x_id)
-    assert ray.worker.global_worker.core_worker.object_exists(x_id)
-
-    # Test wait makes the object local.
-    x_id = a.method.remote()
-    assert not ray.worker.global_worker.core_worker.object_exists(x_id)
-    ok, _ = ray.wait([x_id])
-    assert len(ok) == 1
-    assert ray.worker.global_worker.core_worker.object_exists(x_id)
-
-
-@pytest.mark.skipif(client_test_enabled(), reason="internal api")
-def test_future_resolution_skip_plasma(ray_start_cluster):
-    cluster = ray_start_cluster
-    # Disable worker caching so worker leases are not reused; set object
-    # inlining size threshold so the borrowed ref is inlined.
-    cluster.add_node(
-        num_cpus=1,
-        resources={"pin_head": 1},
-        _system_config={
-            "worker_lease_timeout_milliseconds": 0,
-            "max_direct_call_object_size": 100 * 1024,
-        },
-    )
-    cluster.add_node(num_cpus=1, resources={"pin_worker": 1})
-    ray.init(address=cluster.address)
-
-    @ray.remote(resources={"pin_head": 1})
-    def f(x):
-        return x + 1
-
-    @ray.remote(resources={"pin_worker": 1})
-    def g(x):
-        borrowed_ref = x[0]
-        f_ref = f.remote(borrowed_ref)
-        f_result = ray.get(f_ref)
-        # borrowed_ref should be inlined on future resolution and shouldn't be
-        # in Plasma.
-        assert ray.worker.global_worker.core_worker.object_exists(
-            borrowed_ref, memory_store_only=True
-        )
-        return f_result * 2
-
-    one = f.remote(0)
-    g_ref = g.remote([one])
-    assert ray.get(g_ref) == 4
-
-
-def test_task_output_inline_bytes_limit(ray_start_cluster):
-    cluster = ray_start_cluster
-    # Disable worker caching so worker leases are not reused; set object
-    # inlining size threshold and enable storing of small objects in in-memory
-    # object store so the borrowed ref is inlined.
-    # set task_rpc_inlined_bytes_limit which only allows inline 20 bytes.
-    cluster.add_node(
-        num_cpus=1,
-        resources={"pin_head": 1},
-        _system_config={
-            "worker_lease_timeout_milliseconds": 0,
-            "max_direct_call_object_size": 100 * 1024,
-            "task_rpc_inlined_bytes_limit": 20,
-        },
-    )
-    cluster.add_node(num_cpus=1, resources={"pin_worker": 1})
-    ray.init(address=cluster.address)
-
-    @ray.remote(num_returns=5, resources={"pin_head": 1})
-    def f():
-        return list(range(5))
-
-    @ray.remote(resources={"pin_worker": 1})
-    def sum():
-        numbers = f.remote()
-        result = 0
-        for i, ref in enumerate(numbers):
-            result += ray.get(ref)
-            inlined = ray.worker.global_worker.core_worker.object_exists(
-                ref, memory_store_only=True
-            )
-            if i < 2:
-                assert inlined
-            else:
-                assert not inlined
-        return result
-
-    assert ray.get(sum.remote()) == 10
-
-
-def test_task_arguments_inline_bytes_limit(ray_start_cluster):
-    cluster = ray_start_cluster
-    cluster.add_node(
-        num_cpus=1,
-        resources={"pin_head": 1},
-        _system_config={
-            "max_direct_call_object_size": 100 * 1024,
-            # if task_rpc_inlined_bytes_limit is greater than
-            # max_grpc_message_size, this test fails.
-            "task_rpc_inlined_bytes_limit": 18 * 1024,
-            "max_grpc_message_size": 20 * 1024,
-        },
-    )
-    cluster.add_node(num_cpus=1, resources={"pin_worker": 1})
-    ray.init(address=cluster.address)
-
-    @ray.remote(resources={"pin_worker": 1})
-    def foo(ref1, ref2, ref3):
-        return ref1 == ref2 + ref3
-
-    @ray.remote(resources={"pin_head": 1})
-    def bar():
-        # if the refs are inlined, the test fails.
-        # refs = [ray.put(np.random.rand(1024) for _ in range(3))]
-        # return ray.get(
-        #     foo.remote(refs[0], refs[1], refs[2]))
-
-        return ray.get(
-            foo.remote(
-                np.random.rand(1024),  # 8k
-                np.random.rand(1024),  # 8k
-                np.random.rand(1024),
-            )
-        )  # 8k
-
-    ray.get(bar.remote())
-
-
-# This case tests whether gcs-based actor scheduler works properly with
-# a normal task co-existed.
-def test_schedule_actor_and_normal_task(ray_start_cluster):
-    cluster = ray_start_cluster
-    cluster.add_node(
-        memory=1024 ** 3, _system_config={"gcs_actor_scheduling_enabled": True}
-    )
-    ray.init(address=cluster.address)
-    cluster.wait_for_nodes()
-
-    @ray.remote(memory=600 * 1024 ** 2, num_cpus=0.01)
-    class Foo:
-        def method(self):
-            return 2
-
-    @ray.remote(memory=600 * 1024 ** 2, num_cpus=0.01)
-    def fun(singal1, signal_actor2):
-        signal_actor2.send.remote()
-        ray.get(singal1.wait.remote())
-        return 1
-
-    singal1 = SignalActor.remote()
-    signal2 = SignalActor.remote()
-
-    o1 = fun.remote(singal1, signal2)
-    # Make sure the normal task is executing.
-    ray.get(signal2.wait.remote())
-
-    # The normal task is blocked now.
-    # Try to create actor and make sure this actor is not created for the time
-    # being.
-    foo = Foo.remote()
-    o2 = foo.method.remote()
-    ready_list, remaining_list = ray.wait([o2], timeout=2)
-    assert len(ready_list) == 0 and len(remaining_list) == 1
-
-    # Send a signal to unblock the normal task execution.
-    ray.get(singal1.send.remote())
-
-    # Check the result of normal task.
-    assert ray.get(o1) == 1
-
-    # Make sure the actor is created.
-    assert ray.get(o2) == 2
-
-
-# This case tests whether gcs-based actor scheduler works properly
-# in a large scale.
-def test_schedule_many_actors_and_normal_tasks(ray_start_cluster):
-    cluster = ray_start_cluster
-
-    node_count = 10
-    actor_count = 50
-    each_actor_task_count = 50
-    normal_task_count = 1000
-    node_memory = 2 * 1024 ** 3
-
-    for i in range(node_count):
-        cluster.add_node(
-            memory=node_memory,
-            _system_config={"gcs_actor_scheduling_enabled": True} if i == 0 else {},
-        )
-    ray.init(address=cluster.address)
-    cluster.wait_for_nodes()
-
-    @ray.remote(memory=100 * 1024 ** 2, num_cpus=0.01)
-    class Foo:
-        def method(self):
-            return 2
-
-    @ray.remote(memory=100 * 1024 ** 2, num_cpus=0.01)
-    def fun():
-        return 1
-
-    normal_task_object_list = [fun.remote() for _ in range(normal_task_count)]
-    actor_list = [Foo.remote() for _ in range(actor_count)]
-    actor_object_list = [
-        actor.method.remote()
-        for _ in range(each_actor_task_count)
-        for actor in actor_list
-    ]
-    for object in ray.get(actor_object_list):
-        assert object == 2
-
-    for object in ray.get(normal_task_object_list):
-        assert object == 1
-
-
-# This case tests whether gcs-based actor scheduler distributes actors
-# in a balanced way. By default, it uses the `SPREAD` strategy of
-# gcs resource scheduler.
-@pytest.mark.parametrize("args", [[5, 20], [5, 3]])
-def test_actor_distribution_balance(ray_start_cluster, args):
-    cluster = ray_start_cluster
-
-    node_count = args[0]
-    actor_count = args[1]
-
-    for i in range(node_count):
-        cluster.add_node(
-            memory=1024 ** 3,
-            _system_config={"gcs_actor_scheduling_enabled": True} if i == 0 else {},
-        )
-    ray.init(address=cluster.address)
-    cluster.wait_for_nodes()
-
-    @ray.remote(memory=100 * 1024 ** 2, num_cpus=0.01)
-    class Foo:
-        def method(self):
-            return ray.worker.global_worker.node.unique_id
-
-    actor_distribution = {}
-    actor_list = [Foo.remote() for _ in range(actor_count)]
-    for actor in actor_list:
-        node_id = ray.get(actor.method.remote())
-        if node_id not in actor_distribution.keys():
-            actor_distribution[node_id] = []
-        actor_distribution[node_id].append(actor)
-
-    if node_count >= actor_count:
-        assert len(actor_distribution) == actor_count
-        for node_id, actors in actor_distribution.items():
-            assert len(actors) == 1
-    else:
-        assert len(actor_distribution) == node_count
-        for node_id, actors in actor_distribution.items():
-            assert len(actors) <= int(actor_count / node_count)
-
-
-# This case tests whether RequestWorkerLeaseReply carries normal task resources
-# when the request is rejected (due to resource preemption by normal tasks).
-def test_worker_lease_reply_with_resources(ray_start_cluster):
-    cluster = ray_start_cluster
-    cluster.add_node(
-        memory=2000 * 1024 ** 2,
-        _system_config={
-            "gcs_resource_report_poll_period_ms": 1000000,
-            "gcs_actor_scheduling_enabled": True,
-        },
-    )
-    node2 = cluster.add_node(memory=1000 * 1024 ** 2)
-    ray.init(address=cluster.address)
-    cluster.wait_for_nodes()
-
-    @ray.remote(memory=1500 * 1024 ** 2)
-    def fun(signal):
-        signal.send.remote()
-        time.sleep(30)
-        return 0
-
-    signal = SignalActor.remote()
-    fun.remote(signal)
-    # Make sure that the `fun` is running.
-    ray.get(signal.wait.remote())
-
-    @ray.remote(memory=800 * 1024 ** 2)
-    class Foo:
-        def method(self):
-            return ray.worker.global_worker.node.unique_id
-
-    foo1 = Foo.remote()
-    o1 = foo1.method.remote()
-    ready_list, remaining_list = ray.wait([o1], timeout=10)
-    # If RequestWorkerLeaseReply carries normal task resources,
-    # GCS will then schedule foo1 to node2. Otherwise,
-    # GCS would keep trying to schedule foo1 to
-    # node1 and getting rejected.
-    assert len(ready_list) == 1 and len(remaining_list) == 0
-    assert ray.get(o1) == node2.unique_id
 
 
 if __name__ == "__main__":
